DEBUG

+++ unet3d/dataset.py
@@ -22,7 +22,7 @@ class SegmentationDatasetPersistent(PersistentDataset):
         else:
             keys = ("image", "label")
         transforms.append(LoadImageD(keys=keys, image_only=True, ensure_channel_first=True,
-                                     dtype=(np.float32, int)))
+                                     dtype=(np.float32, np.uint8)))
 
         if crop_foreground:
             foreground_func = partial(percentile_threshold, percentile=foreground_percentile)
